requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
websockets==12.0
mysql-connector-python==9.0.0
Faker==27.0.0
redis==5.0.8
python_version >= "3.7"
psycopg2-binary==2.9.9
